Crate substrate_network_libp2p

Networking layer of Substrate.

Important: This crate is unstable and the API and usage may change.

Re-exports

pub use libp2p::multiaddr;

Modules

identity
A node's network identity keys.

Macros

build_multiaddr
Easy way for a user to create a Multiaddr.

Structs

Multiaddr
Representation of a Multiaddr.

NetworkConfiguration
Network service configuration.

NetworkState
Returns general information about the networking.

NetworkStateNotConnectedPeer

NetworkStatePeer

PeerId
Identifier of a peer of the network.

ProtocolId
Name of a protocol, transmitted on the wire. Should be unique for each chain.

RegisteredProtocol
Connection upgrade for a single protocol.

Service
Network service. Must be polled regularly in order for the networking to work.

Enums

NetworkStatePeerEndpoint

NodeKeyConfig
The configuration of a node's secret key, describing the type of key and how it is obtained. A node's identity keypair is the result of the evaluation of the node key configuration.

NonReservedPeerMode
The policy for connections to non-reserved peers.

ParseErr
Error that can be generated by parse_str_addr.

PublicKey
The public key of a node's identity keypair.

Secret
The configuration options for obtaining a secret key K.

ServiceEvent
Event produced by the service.

Severity
The severity of misbehaviour of a peer that is reported.

Traits

CustomMessage
Implemented on messages that can be sent or received on the network.

Functions

parse_str_addr
Parses a string address and returns the component, if valid.

start_service
Starts the substrate libp2p service.

Type Definitions

Ed25519Secret
The options for obtaining a Ed25519 secret key.

Secp256k1Secret
The options for obtaining a Secp256k1 secret key.