Trait substrate_network_libp2p::CustomMessage

pub trait CustomMessage {
    fn into_bytes(self) -> Vec<u8>;

    fn from_bytes(bytes: &[u8]) -> Result<Self, ()>
    where
        Self: Sized;
}

Implemented on messages that can be sent or received on the network.

Required methods

fn into_bytes(self) -> Vec<u8>

Turns a message into the raw bytes to send over the network.

fn from_bytes(bytes: &[u8]) -> Result<Self, ()> where
Self: Sized,

Tries to parse bytes received from the network into a message.

Loading content...